agent/requirements.txt
websockets>=12.0
python-dotenv>=1.0.0
watchdog>=3.0.0
paramiko>=3.0.0
